% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseacc.R
\docType{package}
\name{gseacc-package}
\alias{gseacc}
\alias{gseacc-package}
\title{gseacc}
\description{
A C++ singel sample GSEA for Rna-seq and scRna-seq datasets.
As main features:
- Mutithreading
- Low RAM usage
}
\details{
Imports
}
